updating element item ProductionPrice

+++ src/Resourcerer.Dtos/V1/Items/V1UpdateElementItem.cs
@@ -4,7 +4,7 @@ public class V1UpdateElementItem : IDto
 {
     public Guid ItemId { get; set; }
     public string? Name { get; set; }
-    public double ProductionPrice { get; set; }
+    public decimal ProductionPrice { get; set; }
     public double ProductionTimeSeconds { get; set; }
     public double? ExpirationTimeSeconds { get; set; }
     public Guid CategoryId { get; set; }
